refactor: split client creation from index module to prevent circle
